OnChallenge modifies response object. Instead, use exception from authentication failure to short-circuit onchallenge handler when failure source is known (used Maskinporten token on Altinn token endpoint), and set the response there instead.
