+++ src/Altinn.Correspondence.API/Auth/MaskinportenSecurityTokenException.cs
@@ -1,0 +1,6 @@
+﻿namespace Altinn.Correspondence.API.Auth
+{
+    public class MaskinportenSecurityTokenException : Exception
+    {
+    }
+}
